Add files via upload
How does Touchless Arduino Doorbell Work?
When someone wants to ring the bell, he/she should put his/her hands in front of the MQ3 sensor after sanitizing hands (2 to 5 cm. distance).

At this moment mq3 generates an analog output which will drive the interfaced Arduino-Uno to generate a signal at pin-8 and pin- 9. 

Here the LED is connected to pin-8 which will glow when mq3 senses the odor of sanitizer.

Pin 9 is connected to ULN2003IC which actually drives the relay at the same time and the relay is connected to the doorbell so it will ring.

Advantages of Automatic Touchless Doorbell
Automatic bell ring facility at the entrance of the Building
Mandatory use of hand- sanitizer so protection From corona